Initial commit.
This is an app made for learning purposes as a part of Londroids Study Jams.

The Study Jams follow Udacity's Android Development for Beginners course. The requirements and materials used are from that course.

The app is made with Firebase to highlight how a fairly functional app can be made with a real-time database in very limited time. The use of Firebase is not in the scope of "Android Development for Beginners" rather the personal interest of the student.