[UPDATE] The project has been updated to Firebase 2.0! There are changes to the build.gradle, importing libraries, etc. The whole endeavour took mere 20 minutes!
This is just preliminary test for the app.

More information about Firebase:
https://www.firebase.com/

Android quick start with Firebase video tutorial:
https://www.youtube.com/watch?v=cab-p7pJBDw

A link with the JSON file for the root of the Firebase project can be found here:
https://drive.google.com/open?id=0B9B86M6T4uiaNkk0YVQ3dGwzajA

This is displaying just the basics.
There is little data - title, image url and body in the database. The title and image url is extracted and displayed in a TextView and ImageView with the help of Glide for the latter.

The code is following the great explanation by Frank van Puffelen to get the data from Firebase:
 http://stackoverflow.com/questions/32108969/why-do-i-get-failed-to-bounce-to-type-when-i-turn-json-from-firebase-into-java

 The example is showing how to load GIFs with the help of Glide (one of the image urls is pointing to .gif in the JSON file).
 Solution for Glide and GIF loading problem:
 http://stackoverflow.com/questions/31082330/show-gif-file-with-glide-image-loading-and-caching-library

 Happy coding :)